Added values to led start and end indexes
